refactor: reduce complexity of some components
- [x] Add zero timeout for playing sounds
- [x] Extract icons from GoogleSTTInput component
- [x] Split GoogleSttChat component into small functions
