+++ src/assets/icons/MicIcon.tsx
@@ -1,0 +1,35 @@
+const MicIcon = ({ color = '#FFFFFF' }) => {
+  return (
+    <svg
+      viewBox='0 0 512.000000 512.000000'
+      className='h-6 w-6'
+      preserveAspectRatio='xMidYMid meet'
+    >
+      <g
+        transform='translate(0.000000,512.000000) scale(0.100000,-0.100000)'
+        fill={color}
+        stroke='none'
+      >
+        <path
+          d='M2420 4784 c-234 -50 -436 -254 -485 -489 -22 -104 -22 -1766 0
+                                    -1870 60 -287 330 -505 625 -505 237 0 462 139 568 350 73 147 72 119 72 1090
+                                    0 971 1 943 -72 1090 -129 257 -422 396 -708 334z'
+        />
+        <path
+          d='M1375 3026 c-37 -17 -70 -52 -84 -89 -17 -45 -14 -596 4 -698 38
+                                    -220 140 -411 307 -576 191 -187 407 -299 686 -354 l112 -22 0 -323 0 -323
+                                    -349 -3 c-335 -3 -351 -4 -377 -24 -53 -39 -69 -71 -69 -134 0 -63 16 -95 69
+                                    -134 27 -21 34 -21 886 -21 852 0 859 0 886 21 53 39 69 71 69 134 0 63 -16
+                                    95 -69 134 -26 20 -42 21 -377 24 l-349 3 0 323 0 324 98 17 c264 47 496 163
+                                    678 337 147 141 243 291 297 463 42 132 48 215 45 537 -3 281 -4 298 -24 324
+                                    -39 53 -71 69 -134 69 -63 0 -95 -16 -134 -69 -20 -26 -21 -44 -27 -359 -5
+                                    -326 -6 -334 -32 -412 -37 -110 -89 -195 -176 -288 -225 -237 -579 -349 -927
+                                    -292 -237 38 -427 136 -579 297 -84 88 -136 175 -172 283 -26 78 -27 86 -32
+                                    412 -6 315 -7 333 -27 359 -11 15 -32 37 -46 47 -34 25 -113 32 -153 13z'
+        />
+      </g>
+    </svg>
+  );
+};
+
+export default MicIcon;
