+++ src/assets/icons/MicWaveIcon.tsx
@@ -1,0 +1,77 @@
+const MicWaveIcon = () => {
+  return (
+    <svg
+      viewBox='0 0 512.000000 512.000000'
+      className='h-6 w-6'
+      preserveAspectRatio='xMidYMid meet'
+    >
+      <g
+        transform='translate(0.000000,512.000000) scale(0.100000,-0.100000)'
+        fill='#FFFFFF'
+        stroke='none'
+      >
+        <path
+          d='M2410 4945 c-283 -63 -487 -250 -581 -530 -23 -70 -24 -72 -27 -662
+                                      l-3 -593 761 0 761 0 -3 593 c-3 589 -4 592 -27 662 -50 151 -133 275 -248
+                                      369 -75 62 -198 127 -288 151 -85 23 -264 28 -345 10z'
+        />
+        <path
+          d='M543 4769 c-44 -13 -179 -159 -264 -287 -417 -623 -359 -1455 141
+                                      -2013 77 -87 112 -109 168 -109 100 0 168 84 147 180 -4 17 -32 59 -62 93
+                                      -181 201 -289 407 -345 658 -30 137 -32 397 -5 529 54 254 162 469 336 663 39
+                                      44 73 92 76 109 16 73 -33 158 -103 178 -42 11 -46 11 -89 -1z'
+        />
+        <path
+          d='M4486 4770 c-41 -13 -85 -58 -98 -101 -19 -64 -4 -102 73 -187 171
+                                      -188 282 -407 336 -662 27 -132 25 -392 -5 -529 -54 -239 -152 -435 -309 -615
+                                      -37 -43 -76 -92 -85 -109 -51 -92 22 -207 131 -207 60 0 93 21 171 109 223
+                                      248 361 551 405 888 63 478 -93 972 -415 1315 -95 102 -131 119 -204 98z'
+        />
+        <path
+          d='M965 4336 c-60 -26 -181 -183 -244 -317 -69 -143 -92 -243 -98 -409
+                                      -9 -246 43 -437 170 -625 96 -142 168 -205 237 -205 50 0 107 35 130 80 37 73
+                                      24 123 -59 217 -65 74 -118 166 -148 258 -35 108 -43 278 -18 392 23 107 91
+                                      243 164 328 67 77 86 120 78 171 -16 96 -120 150 -212 110z'
+        />
+        <path
+          d='M4024 4336 c-62 -28 -97 -111 -75 -177 6 -19 38 -65 71 -104 74 -85
+                                      133 -201 161 -314 18 -70 21 -105 17 -205 -8 -187 -62 -326 -179 -459 -83 -94
+                                      -96 -144 -59 -217 13 -26 34 -47 60 -60 85 -43 149 -17 242 99 119 149 183
+                                      283 220 461 17 79 19 121 15 245 -6 178 -29 273 -106 430 -49 100 -160 249
+                                      -213 288 -34 25 -113 32 -154 13z'
+        />
+        <path
+          d='M1403 3919 c-41 -12 -106 -86 -146 -168 -30 -63 -32 -72 -32 -181 1
+                                      -132 18 -189 82 -274 48 -63 92 -89 148 -88 132 2 191 143 105 252 -26 33 -35
+                                      54 -38 96 -4 50 -2 59 37 116 33 50 41 70 41 107 0 61 -29 107 -83 131 -46 21
+                                      -66 23 -114 9z'
+        />
+        <path
+          d='M3603 3910 c-54 -24 -83 -70 -83 -131 0 -37 8 -57 41 -107 39 -57 41
+                                      -66 37 -116 -3 -42 -12 -63 -38 -96 -86 -108 -27 -250 104 -252 59 -1 100 24
+                                      151 94 64 86 89 172 83 291 -3 78 -9 102 -36 158 -42 87 -103 155 -150 169
+                                      -50 13 -60 13 -109 -10z'
+        />
+        <path
+          d='M1802 2568 c4 -271 6 -298 27 -363 68 -209 207 -372 395 -464 122
+                                      -60 190 -75 336 -75 146 0 214 15 336 75 188 92 327 255 395 464 21 65 23 92
+                                      27 363 l4 292 -762 0 -762 0 4 -292z'
+        />
+        <path
+          d='M1278 2539 c-24 -12 -46 -35 -59 -61 -20 -38 -21 -53 -15 -145 29
+                                      -475 306 -893 741 -1117 108 -56 270 -110 383 -127 l82 -13 0 -308 0 -308
+                                      -245 0 c-227 0 -249 -2 -285 -20 -45 -23 -80 -80 -80 -130 0 -50 35 -107 80
+                                      -130 38 -19 58 -20 680 -20 622 0 642 1 680 20 45 23 80 80 80 130 0 50 -35
+                                      107 -80 130 -36 18 -58 20 -285 20 l-245 0 0 308 0 308 82 13 c252 39 511 170
+                                      709 359 121 114 207 232 279 377 87 178 125 319 136 508 6 92 5 107 -15 145
+                                      -38 75 -124 101 -201 62 -52 -26 -80 -81 -80 -153 0 -71 -25 -215 -52 -297
+                                      -91 -280 -320 -528 -596 -645 -150 -63 -238 -80 -417 -79 -174 1 -232 12 -390
+                                      73 -227 87 -441 285 -555 511 -65 129 -110 307 -110 438 0 133 -113 209 -222
+                                      151z'
+        />
+      </g>
+    </svg>
+  );
+};
+
+export default MicWaveIcon;
