+++ src/components/atoms/SettingsDropdown.js
@@ -74,7 +74,7 @@ export default function SettingsDropdown({
       </Transition>
       <Popover.Button
         ref={setReferenceElement}
-        className='inline-flex w-full justify-center rounded-md bg-opacity-20 px-4 py-2 text-sm font-medium text-black hover:bg-opacity-30 focus:outline-none focus-visible:ring-2 focus-visible:ring-white focus-visible:ring-opacity-75'
+        className='inline-flex w-full justify-center rounded-md bg-opacity-20 px-4 py-2 text-sm font-medium text-black hover:bg-opacity-30 focus:outline-none focus-visible:ring-2 focus-visible:ring-white focus-visible:ring-opacity-75 outline-none'
       >
         <SettingIcon />
       </Popover.Button>
